Refactoring to split up the monolithic pdfstitcher.py
