+++ pdfstitcher/ui/dialogs.py
